migoinfer: nilchan support.
Add support for use of nilchan.
If a channel is undeclared and used by send/recv, make an artificial channel.
If a nil channel is being passed to function, make an artificial channel in the
caller. (Nilchan is handled after return value handling, so will only add real
nilchans).
If a raw nilchan (variables identified as nil:chan constant) is found, they are
given a numbered name to make sure they block independently.